fix(kanban): enforce lostReasonRequired on drag-to-lost-stage (closes #4447)
When dragging a lead to a pipeline stage with isLostStage=true, or dropping
it on the bottom "Lost" zone, show a LostReasonDialog before calling the
backend — matching the behavior already in the lead detail page. Extracts
LostReasonDialog to a shared component used by both the detail page and the
Kanban page.
