+++ src/routes/_app/_auth/dashboard/_layout.leads.$leadId.lazy.tsx
@@ -24,8 +24,7 @@ import {
   useSupabaseRelationshipsByEntity,
   type MappedRelationship,
 } from "@/hooks/use-supabase-relationships";
-import { useSupabaseLostReasonsList } from "@/hooks/use-supabase-lost-reasons";
-import { useSupabaseOrgSettings } from "@/hooks/use-supabase-organizations";
+import { LostReasonDialog } from "@/components/crm/lost-reason-dialog";
 import { useTagDefinitions } from "@/hooks/use-tag-definitions";
 import { useCategoryDefinitions } from "@/hooks/use-category-definitions";
 import { supabaseKeys } from "@/lib/supabase/query-keys";
@@ -50,7 +49,6 @@ import { activitiesToFeedEntries } from "@/components/crm/activity-feed-adapter"
 import { ScheduledActivitiesList } from "@/components/shared/scheduled-activities-list";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
 import {
   Dialog,
   DialogContent,
@@ -58,13 +56,6 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { RichTextEditor, plateJsonToText } from "@/components/gabinet/rich-text-editor";
 import {
   DropdownMenu,
@@ -149,90 +140,6 @@ function PipelineProgressBar({
         );
       })}
     </div>
-  );
-}
-
-// --- LostReasonDialog ---
-
-function LostReasonDialog({
-  open,
-  onOpenChange,
-  onConfirm,
-  organizationId,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  onConfirm: (reason: string) => void;
-  organizationId: string;
-}) {
-  const { data: lostReasons } = useSupabaseLostReasonsList(organizationId);
-  const { data: orgSettings } = useSupabaseOrgSettings(organizationId);
-  const { t } = useTranslation();
-  const [selectedReason, setSelectedReason] = useState("");
-  const [customReason, setCustomReason] = useState("");
-
-  // Hide freeform custom entry when org has disabled custom lost reasons.
-  // Defaults to allowed (fail-open) when settings are not yet loaded or not configured.
-  const allowCustom = orgSettings?.allowCustomLostReason !== false;
-
-  const handleSubmit = () => {
-    const reason = selectedReason === "__custom__" ? customReason : selectedReason;
-    if (reason.trim()) {
-      onConfirm(reason.trim());
-      setSelectedReason("");
-      setCustomReason("");
-    }
-  };
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="sm:max-w-[400px]">
-        <DialogHeader>
-          <DialogTitle>{t('detail.lostDialog.title')}</DialogTitle>
-        </DialogHeader>
-        <div className="space-y-3">
-          <Select value={selectedReason} onValueChange={setSelectedReason}>
-            <SelectTrigger>
-              <SelectValue placeholder={t('detail.lostDialog.selectReason')} />
-            </SelectTrigger>
-            <SelectContent>
-              {lostReasons?.map((r) => (
-                <SelectItem key={r._id} value={r.label}>
-                  {r.label}
-                </SelectItem>
-              ))}
-              {allowCustom && (
-                <SelectItem value="__custom__">{t('detail.lostDialog.customReason')}</SelectItem>
-              )}
-            </SelectContent>
-          </Select>
-          {selectedReason === "__custom__" && allowCustom && (
-            <Input
-              value={customReason}
-              onChange={(e) => setCustomReason(e.target.value)}
-              placeholder={t('detail.lostDialog.customPlaceholder')}
-              autoFocus
-            />
-          )}
-        </div>
-        <DialogFooter>
-          <Button variant="outline" size="sm" onClick={() => onOpenChange(false)}>
-            {t('common.cancel')}
-          </Button>
-          <Button
-            size="sm"
-            variant="destructive"
-            disabled={
-              !selectedReason ||
-              (selectedReason === "__custom__" && !customReason.trim())
-            }
-            onClick={handleSubmit}
-          >
-            {t('detail.lostDialog.confirm')}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
   );
 }
